Python/DjangoTests/Loja/requirements.txt
beautifulsoup4==4.9.0
Django==2.2.9
django-bootstrap4==1.1.1
django-stdimage==5.1.1
Pillow==7.1.1
progressbar2==3.50.1
python-utils==2.4.0
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
